Added- Lists, Contacts Section, and "Nav Bar"
I added lists under a few elements of the Timeline in order to add more information. Also added some paragraphs. The Contacts section at the bottom works including a mail-to action that allows the user to open the mail app. I also added a Nav Bar that links onto the page itself to make it easier to navigate the page to each section.

In the future I will:
* add images
* do information to all the elements.
* create a grid for projects to display nice links + images