Added styling to navbar
I added styling to the navbar. It has an image for the logo as well as some links to points on the page. It has a black background, white text (vertically centered) with a 20px margin to the right. The display is inline - block.

Future:
1. Figure out exactly the design for the page
2. Update Nav bar and figure out how to properly size the elements to the left (text links)
   a. On that thought I think the easy path might be create a wrapping div with an associated class that handles the text links separately.
   b. I can also try doing some CSS fun stuff to get all list elements that do not contain a link with an image.
3. The next part is creating a nice entry video that passively runs in the background. That is a later goal. For now a nice image as a placeholder. The idea is have a short 30 second video displaying what I am about with short snips of important/cool events in my life so that people coming to my page understnad what I am about.

Notes:
I am taking inspiration from the following 2 links:
1.https://benscott.dev/
2.https://www.benrogers.dev/
These two websites have a lot of things I would like to implement in my portfolio webpage in some capacity